feat(pipelines): support timeout in CodeBuildStep (#17351)
Make sure we can set timeout for those long running tests.

There's one thing I didn't get. In `_codebuild-factory.ts/produceAction()` I expected that this line:

```
    const projectOptions = mergeCodeBuildOptions(options.codeBuildDefaults, this.props.projectOptions);
```

would merge in the timeout if set. But it didn't do that. So that's why I pull the timeout in the later `codebuild.PipelineProject` explicitly via:

```
      timeout: this.props.projectOptions?.timeout,
```

Not sure why I had to do that. I noticed that `options.codeBuildDefaults` didn't have the timeout field, perhaps that's why the merge didn't work, or perhaps I don't understand this merge, and what I did was right :-)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*