chore(cloud-assembly-schema): expose lookupRole on the cloudformation stack artifact
This PR exposes information on the bootstrap lookup role on the
CloudFormation stack artifact. This enables the CLI to assume the lookup
role during cli operations in order to lookup information in the stack
account.

Along with the ARN of the lookup role, this also exposes a
`requiresBootstrapStackVersion` property which is set to `8` (the
version the lookup role was given ReadOnlyAccess), and the
`bootstrapStackVersionSsmParameter` which is needed to lookup the
bootstrap version if a user has renamed the bootstrap stack.

This allows us to first check whether the lookupRole exists and has the
correct permissions prior to using it.