fix(aws-kinesis): remove default shard count when stream mode is on-demand and set default mode to provisioned (#18221)
Change the default Kinesis Data Stream's stream mode to provisioned from undefined to make the active configuration more explicit in the resulting CloudFormation templates.

Fix an issue whereby the shard count is always set when the stream mode is set to on-demand, which is invalid. Shard count still defaults to `1` in provisioned mode, but is left undefined in on-demand mode.

Add validation for the above so that an error is thrown from CDK when specifying on-demand mode with a shard count.

Fixes #18139

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*