fix(appmesh): allow a Virtual Node have as a backend a Virtual Service whose provider is that Node (#18265)
Addresses a circular dependency issue between Virtual Nodes and Virtual Services that works for Virtual Services created with a defined `virtualServiceName` and a randomly generated name. 

One such example of this problem was a case where a Virtual Node had a backend that is a Virtual Service whose provider was given as the same Virtual Node. This led to the Virtual Node being dependent on the creation of the Virtual Service, and the Virtual Service being dependent on the creation of the Virtual Node.

Fixes #17322

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*