Make TagOptions an L2 and wrapped construct
We are moving to make TagOptions there own L2 like construct, previously they were made only as an interface
and added to a resource, but since they are global resources, coupling them to an association doesn't make much sense.

I'm not too sure if this implementation makes that much sense either, I could see how getting rid of the TagOption class
is fine, since it doesn't provide much benefit over the L1, and just use TagOptions class as the sort of L2 for it.

Testing done
------------------
*

Related items
-------------------
* SIM/auto-cut ticket