fix(aws-lambda-event-sources): unsupported properties for SelfManagedKafkaEventSource and ManagedKafkaEventSource (#17965)
This PR fixes a bug in the CDK where some `kafkaEventSource` properties are actually unsupported. These properties exist only for kinesis and dynamodb streams. The existing KafkaEventSourceProps Interface erroneously extends an interface that includes kinesis and dynamodb specific properties. This PR separates these properties into a `Base` interface with shared stream properties for all 3, as well as an interface for `kinesis` and `dynamodb` specific properties. 

Unit testing unavailable because the scope of the PR is to remove properties. It is enough to ensure that current tests still succeed.

We are allowing the breaking changes specified in  `allowed-breaking-changes.txt` because they never worked in the first place.

Fixes #17934.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*