feat(cli): diff now uses the lookup Role for new-style synthesis (#18277)
This PR exposes information on the bootstrap lookup role on the
CloudFormation stack artifact. This enables the CLI to assume the lookup
role during cli operations in order to lookup information in the stack
account.

Along with the ARN of the lookup role, this also exposes a
`requiresBootstrapStackVersion` property which is set to `8` (the
version the lookup role was given ReadOnlyAccess), and the
`bootstrapStackVersionSsmParameter` which is needed to lookup the
bootstrap version if a user has renamed the bootstrap stack.

This allows us to first check whether the lookupRole exists and has the
correct permissions prior to using it.

This also updates the `diff` capability in the CLI (run as part of `cdk diff` or `cdk deploy`)
to use this new functionality. It now will try to assume the lookupRole and if it doesn't exist or
if the bootstrap stack version is not valid, then it will fallback to using the deployRole (what it uses
currently).

This PR also updates the `forEnvironment` function to return whether or not it is returning the
default credentials. This allows the calling function to decide whether or not it actually wants
to use the default credentials.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*