chore: skip post_build on build failure
According to the docs on CodeBuild [build transitions], the `POST_BUILD` step is
always run after `BUILD`, even if `BUILD` fails. This can mean in our pipeline
we can fail the build, and then wait for the (currently very expensive) packing
step to run completely before the job fails. This change short-circuits the
logic so we only run pack if the build succeeded.

[build transitions]: https://docs.aws.amazon.com/codebuild/latest/userguide/view-build-details.html