feat(aws-s3): support number of newer versions to retain in lifecycle policy (#18225)
The AWS Management console and CloudFormation support to set a number of noncurrent versions to be retained for S3 lifecycle transitions as described in #17996. 

This pull request introduces the property for S3 lifecycle configurations as an optional property to CDK. 

The maximum supported number of noncurrent versions that could be retained is 100 which is also documented for the new property. However, no additional check is inserted if the number set by the developer is actually between 0 and 100. If this check is desired, a discussion would be good on how to achieve throwing an error as I don't see a smooth solution. 

Resolves #17996.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*