chore(cli): bootstrap stack downgrading is no longer an error (#18378)
Previously, bootstrap downgrading is an error. That is, when the
bootstrap stack is currently version '9', and you're trying to run `cdk
bootstrap` with version '8', the command will currently fail.

This makes it hard to script around, and in fact this is causing us
problems in the pipeline where we unconditionally run `cdk bootstrap`
and an accidentally upgraded bootstrap stack fails all tests.

This is happening for no good reason, since the bootstrap stacks are
backwards compatible.

Turn this case from an error into a successful no-op.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*