fix(pipelines): "Maximum schema version supported" error
We are currently releasing both v1 and v2 streams of the CDK in
parallel, with changes landing on v1 before they land on v2. By default,
CDK Pipelines will use the "latest" CLI version.

However, because "latest" always comes from the v2 version stream (since
that has the highest number), there may be changes to the cx protocol in
a v1 app that aren't supported by the v2 CLI yet.

The "correct" solution to this would be one of (a) releasing v1 and
v2 synchronously and making sure both have the same change set; or (b)
not having a v2 CLI at all. Both of these would require significant
engineering effort to resolve though.

In the mean time, add the concept of a "preferred CLI version" to CDK
Pipelines, set to `1` for most packages but to `2` for `aws-cdk-lib`,
and have NPM install the "latest" CLI from the same version stream
that the library is from.

Fixes #18370.