chore: fix CloudFormation spec generation
Two issues with the latest (v52.0.0) CloudFormation spec generation. First, we
are still combining "V2" libraries into single modules. This means new
generation services (like Inspector V2) are bundled into the existing (e.g.,
`@aws-cdk/aws-inspector`) modules.

Second, there was an issue with the Lightsail spec due to one of the properties
being "GetObject", which is not jsii-compliant. Introducing a hack here to add
an underscore in front of the Typescript property name (but not the generated
CloudFormation). Very open to better ideas, but this does work.