chore: fix CloudFormation spec generation (#18441)
Two issues with the latest (v52.0.0) CloudFormation spec generation. First, we
are still combining "V2" libraries into single modules. This means new
generation services (like Inspector V2) are bundled into the existing (e.g.,
`@aws-cdk/aws-inspector`) modules.

Second, there was an issue with the Lightsail spec due to one of the properties
being "GetObject", which is not jsii-compliant. Introducing a hack here to add
an underscore in front of the Typescript property name (but not the generated
CloudFormation). Very open to better ideas, but this does work.

Once this has been pushed, we can regenerate the spec (now on version v53.0.0) as usual.

related #18305

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*