fix(cli): warning to upgrade to bootstrap version >= undefined
When we try to assume the lookup role and fail, we print out a warning
telling the user to upgrade to the required bootstrap stack version.

In cases where the user is not using the default stack synthesizer
(legacy, custom, etc.) the cli might not have information on the lookup
role. In those cases where we don't have the necessary information on
the ARN of the lookup role or the required bootstrap version, we will
not print any warnings.