fix(secretsmanager): Secret requires KMS key for some same-account access (#17812)
Fix for #15450 Previous code did not check if the account IDs were the different. This checks if CDK is able to resolve the account ids and they are different then fail otherwise let the user create a secret.

FYI first PR. Let me know if there is something that I missed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*