fix(pipelines): graphnode dependencies can have duplicates (#18450)
`GraphNode.allDeps` allows duplicate dependencies to be returned. This does not have any affect on the performance of the pipelines module, but looks ugly. This was noticed in https://github.com/cdklabs/cdk-pipelines-github/pull/67, where the dependencies are written out to the `deploy.yaml` file. 

I did not change the underlying `GraphNode.dependencies` structure to be a set (although I think it should) because I feel like that is a breaking change. So instead I've preserved the structure of the API and deduplicated the list of GraphNodes.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*