fix(elasticloadbalancingv2): ApplicationLoadBalancer.logAccessLogs does not grant all necessary permissions
`ALB.logAccessLogs` today grants the ELB account Put/Abort on the bucket. The
NLB code extends this to also grant permissions to the
`delivery.logs.amazonaws.com` service principal. The ALB documentation now
states that the permissions required for ALB are the same as NLB, so
consolidating the code back into the base.

fixes #18367