fix: dependency cycles
We happily generate templates with dependency cycles. Those would be
undeployable in CloudFormation, so we should make it impossible that
templates like these pass unit test validation.

We could make it a separate assertion method
(`Template.fromStack().assertNoCycles()`), but the only thing that will
do is give you an opportunity to forget to put the test in.

Instead, we just check it by default for every generated template.