chore(cfnspec): outdated spec complaints on every run (#18689)
We recently changed the cfnspec bump to always succeed, and simply
record which specs have been held back because they contain errors.

Except we didn't do it completely right, and now it complains on
every single run about specs that are outdated, instead of only
when the spec actually gets bumped.

Fix that.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*