chore: optimize autolabeler (#18702)
The autolabeler will now do different things depending on how the issue comes in. If it's guidance, it's automatically assigned to OSDS. If it's a PR, the autolabeler will not continue to work if the PR has `contribution/core`. For issues, the autolabeler will not continue to work on the issue if `needs-triage` is removed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*