feat(cloudfront-origins): extend `readTimeout` maximum value for `HttpOriginProps` (#18697)
This pull request extends the maximum value of the `readTimeout` property for `HttpOriginProps` up to 180 seconds. This allows the use of the construct also in the case that a limit increase has been approved for CloudFront [origin request timeout quota](https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/cloudfront-limits.html#limits-web-distributions).

The 180 seconds cap should be AWS hard limit for that quota.

Closes #18628

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*