fix(ec2): UserData.addSignalOnExitCommand uses invalid resource logicalID
If both `addSignalOnExitCommand` _and_ `userDataCausesReplacement` are
both used it results in an invalid logicalId being used in the
`cfn-signal` call. This is due to `addSignalOnExitCommand` getting the
logicalID from `Stack.getLogicalId` which does not take into
consideration logicalId overrides which `userDataCausesReplacement`
uses.

This updates `addSignalOnExitCommand` to use the `logicalId` of the
resource.

fixes #12749