feat(ecs): expose image name in container definition (#17793)
When adding a container to a task definition it is currently impossible to retrieve
its image name because it's the container definition that is responsible for binding
the `AssetImage`.

Expose the `imageName` so that it's possible to reference it in another construct
that needs to use/depend on the asset hash. Example: a custom resource running
a Fargate task that needs to rerun if the underlying container image changes but
not necessarily when the task definition changes.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*