feat(ec2): support KMS keys for block device mappings for both instances and launch templates (#18326)
This pullrequest will add the optional ```kmsKeyId``` property to the ```EbsDeviceOptions``` Interface. Whit this, it will be possible to specify the kmsKeyId used for encrypting the ebs volumes when launching instances. At the moment I already use this via an escape hatch in my projects, but it's not that handy as the block device mapping is an array.

I don't like to only specify a kmsKeyId (= ARN) but accepting an ```kms.IKey``` in the properties would be a bigger change.

fixes: #18309
*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*