chore(lambda-layer-awscli): add update mechanism for AWS CLI (#18780)
We bundle the AWS CLI. This adds a mechanism to keep up with new updates
of the CLI. It adds a script that should be executed periodically to
create a PR with a version update.

We had to install and bump to Python v3 because newer versions of the
AWS CLI have ended support for Python 2.7.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*