chore(lambda-layer-awscli): install awscli with pip and requirements.txt
We are currently bundling the AWS CLI in the Lambda Layer and need a
mechanism to automatically upgrade when new versions are released.

This changes the way we bundle the AWS CLI, from using the install
script to using pip to install from a `requirements.txt` file. This will
enable dependabot to auto upgrade the version of the awscli specified in
requirements.txt