fix(ecr-assets) fix loading image tarballs with existing tags
The current logic for pushing tarball images fails if the tarball being
loaded uses a repository and tag that's already loaded into the docker
daemon. This is due to the way the ecr-assets module parses the output
of the docker load command. If the repository/tag combination already
exists in the daemon, it outputs a message about renaming it, which
breaks the `sed` command parsing its output.

This change updates the `sed` command used for extracting the
repository/tag for tarball images to make it more robust and will now
successfully parse docker load output in the case where the
repository/tag already exist.

fixes #18822