chore(lambda-layer-awscli): install awscli with pip and requirements.txt (#18800)
We are currently bundling the AWS CLI in the Lambda Layer and need a
mechanism to automatically upgrade when new versions are released.

This changes the way we bundle the AWS CLI, from using the install
script to using pip to install from a `requirements.txt` file. This will
enable dependabot to auto upgrade the version of the awscli specified in
requirements.txt

Also added integration tests to ensure that the AWS CLI is executable from within
various Python runtimes. 

The `assetHash` has been updated to fingerprint the entire `layer` directory since
we want to deploy a new version anytime anything in that directory changes, not just the
Dockerfile.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*