chore: allow dependabot to manage pip upgrades for lambda-layer-awscli
This PR follows up #18800 by turning on dependabot updates for `awscli`
in `lambda-layer-awscli`.

Currently the only other place we have pip dependencies are in our tests
and the init templates. dependabot does currently look for security
updates in all of our pip dependencies, this just turns on normal
updates for this specific module.

Need to follow up this PR by adding this process to the PR checklist and
the contributing guide.