chore(cxapi): reduce merge conflicts in feature flags (#18411)
Try to reduce (future) merge conflicts when feature flags are added.

It currently looks like `FUTURE_FLAGS_DEFAULTS` should contain all
flags, and they're all set to `false`. On the `v2-main` branch, they're
all set to `true`.

When adding a new flag, you want to follow suit and add a new line with
the flag default set to `false`; then you get a merge conflict when
merging to `v2-main` branch because all preceding lines will have
changed. The merge conflict alone is annoying, and you'll also be
tempted to put in `true` there, which would be incorrect, and be a
breaking change of behavior.

Instead, this PR gets rid of the entire set of `FUTURE_FLAGS_DEFAULTS`
set to `false` -- there's no point to having them anyway, and it
gets rid of the associated merge conflicts.

Also shore up the docs for these flags a little.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*