fix(aws-cdk): escape spaces while invoking dotnet command
When dotnet command is invoked, if arguments are given
with spaces unescaped, it interprets that as the start
of second command.

In this case, when csharp project is initialized,
`dotnet sln` command is invoked with slnPath and fsprojPath.

Spaces in `slnPath` and `fsprojPath` should be escaped
before providing them as argument to dotnet command

Fixes https://github.com/aws/aws-cdk/issues/18803

----

*By submitting this pull request, I confirm that my contribution is made under the terms of t
he Apache-2.0 license*