chore(assertions): partially revert #18948 (#18998)
Some tests are failing in v2 due to an oversight that had tests depend on user-supplied context. This PR reverts a change made in #18948 that redacted the stack trace for `findXxx()` APIs. Since this is something the user can configure through context values, it is unnecessary. As such, a few tests are now irrelevant.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*