chore: compat checks fail if a release has happened
If a PR is branched off of a version that introduces a new
package, and lives longer than the release, the "does package
exist?" check is wrong, and fails the PR build.

This is usually fixed by another "Update from master", but that's
an annoying step that sometimes takes a while.

- Let's say the branch was forked at 1.142.0
- The branch adds a new package
- That package is released as its initial version as 1.143.0
- When the PR next builds (still at 1.142.0), it will do the following
  checks:
  - Does the package exist at all? (Answer: yes)
  - If so, try to install it at 1.142.0, because that's what we're
    diffing against (This explodes, because the package does not exist
    at that version).

Instead of doing a versionless "does the package exist at all" check,
always check the existence of the actual version we'll be diffing
against.